requirements.txt
get-docker-secret==1.0.2
discord.py==2.2.3
python-dotenv==1.0.0
PyYAML==6.0